build/c: Separate compat and graph objects
The files for graph and compat are distinct from the rest of the code.
This is evident from the Makefile which lists those objects separately
for re-use between the tig binary and the test-graph binary.

Pull them into their own objects so that those files are not recompiled
without need.